requirements.txt
requests
pyyaml
pyyaml-include
jinja2
bs4
pillow